[FEATURE] Preset mysql server on pma login screen
Preset the default hostname of the mysql container by using
the PMA_HOSTS environment variable.
This avoids manually entering the server ('mysql' by default)
on the phpmyadmin login screen.
There will be a selectbox on the login screen to choose a server from,
if multiple servers are given (comma separated) to the PMA_HOSTS env var.